backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
Pillow==10.3.0
python-decouple==3.8
